Changing the security groups
