--- package/cloudshell/cp/aws/domain/services/security_groups_service/aws_security_group_service.py
@@ -1,101 +1,0 @@
-import re
-import uuid
-
-from cloudshell.cp.aws.domain.services.model_parser.port_group_attribute_parser import PortGroupAttributeParser
-from cloudshell.cp.aws.device_access_layer.aws_api import AWSApi
-
-
-class AWSSecurityGroupService(object):
-    QUALI_SECURITY_GROUP = "Quali_security_group"
-    QUALI_SECURITY_GROUP_DESCRIPTION = "Quali Security Group"
-
-    def __init__(self, aws_api):
-        """
-
-        :param AWSApi aws_api:
-        :return:
-        """
-        self.aws_api = aws_api
-
-    def create_security_group(self, ami_deployment_model, aws_ec2_cp_resource_model, ec2_session):
-
-        inbound_ports = PortGroupAttributeParser.parse_port_group_attribute(ami_deployment_model.inbound_ports)
-        outbound_ports = PortGroupAttributeParser.parse_port_group_attribute(ami_deployment_model.outbound_ports)
-        security_group_id = None
-
-        # if the deployment model contains inbound / outbound ports
-        if inbound_ports or outbound_ports:
-            # create a new security port group based on the attributes
-            # in the end creates a tag "CreatedBy : Quali"
-            security_group = self._create_security_group_with_port_group(ec2_session,
-                                                                         inbound_ports,
-                                                                         outbound_ports,
-                                                                         aws_ec2_cp_resource_model.vpc)
-            security_group_id = security_group.group_id
-
-        return security_group_id
-
-    def _create_security_group_with_port_group(self, ec2_session, inbound_ports, outbound_ports, vpc):
-
-        security_group_name = AWSSecurityGroupService.QUALI_SECURITY_GROUP + " " + str(uuid.uuid4())
-
-        # creating the security group
-        security_group = self.aws_api.create_security_group(ec2_session,
-                                                            security_group_name,
-                                                            AWSSecurityGroupService.QUALI_SECURITY_GROUP_DESCRIPTION,
-                                                            vpc)
-
-        # setting tags on the created security group
-        self.TagManagerService.set_security_group_tags(security_group, security_group_name)
-
-        # creating inbound and outbound rules
-        self.creating_port_groups(inbound_ports, outbound_ports, security_group)
-
-        return security_group
-
-    def creating_port_groups(self, inbound_ports, outbound_ports, security_group):
-
-        # adding inbound port rules
-        if inbound_ports:
-            ip_permissions = [self._get_ip_permission_object(port) for port in inbound_ports if port is not None]
-            security_group.authorize_ingress(IpPermissions=ip_permissions)
-
-        # adding outbound port rules
-        if outbound_ports:
-            ip_permissions = [self._get_ip_permission_object(port) for port in outbound_ports if port is not None]
-            security_group.authorize_egress(IpPermissions=ip_permissions)
-
-    @staticmethod
-    def _get_ip_permission_object(port_data):
-        return {
-            'IpProtocol': port_data.protocol,
-            'FromPort': int(port_data.from_port),
-            'ToPort': int(port_data.to_port),
-            'IpRanges': [
-                {
-                    'CidrIp': port_data.destination
-                }
-            ]}
-
-
-class TagManagerService(object):
-    @staticmethod
-    def set_ami_instance_tag(ec2_session, instance, tags):
-        return ec2_session.create_tags(Resources=[instance.id], Tags=tags)
-
-    @staticmethod
-    def set_security_group_tags(security_group, name):
-        return security_group.create_tags(Tags=TagManagerService.get_default_tags(name))
-
-    @staticmethod
-    def get_default_tags(name):
-        return [TagManagerService._get_kvp("Name", name),
-                TagManagerService._get_created_by_kvp()]
-
-    @staticmethod
-    def _get_created_by_kvp():
-        return TagManagerService._get_kvp('CreatedBy', 'Quali')
-
-    @staticmethod
-    def _get_kvp(key, value):
-        return {'Key': key, 'Value': value}
